refactor: FareStrategy를 인터페이스로 추출
