+++ src/main/java/wooteco/subway/domain/fare/strategy/DefaultFareStrategy.java
@@ -1,6 +1,6 @@
-package wooteco.subway.util;
+package wooteco.subway.domain.fare.strategy;
 
-public class FareCalculator {
+public class DefaultFareStrategy implements FareStrategy {
 
     private static final int MINIMUM_FARE = 1250;
     private static final int MINIMUM_ADDITIONAL_FARE = 100;
@@ -10,14 +10,13 @@ public class FareCalculator {
     private static final int SECOND_ADDITIONAL_FARE_DISTANCE = 50;
     private static final int SECOND_ADDITIONAL_STANDARD_DISTANCE = 8;
 
-    private FareCalculator() {
-    }
+    private static final int DEFAULT_PORTION = 1;
 
-    public static int calculate(double distance) {
+    public int calculate(double distance) {
         return MINIMUM_FARE + addOverTen(distance) + addOverFifty(distance);
     }
 
-    private static int addOverTen(double distance) {
+    private int addOverTen(double distance) {
         if (isUnderAdditionalFareDistance(distance, FIRST_ADDITIONAL_FARE_DISTANCE)) {
             return 0;
         }
@@ -28,7 +27,7 @@ public class FareCalculator {
         return calculateFare(portion);
     }
 
-    private static int addOverFifty(double distance) {
+    private int addOverFifty(double distance) {
         if (isUnderAdditionalFareDistance(distance, SECOND_ADDITIONAL_FARE_DISTANCE)) {
             return 0;
         }
@@ -36,19 +35,19 @@ public class FareCalculator {
         return calculateFare(portion);
     }
 
-    private static boolean isUnderAdditionalFareDistance(double distance, int additionalFareDistance) {
+    private boolean isUnderAdditionalFareDistance(double distance, int additionalFareDistance) {
         return distance <= additionalFareDistance;
     }
 
-    private static int getPortion(double distance, int additionalFareDistance, int additionalStandardDistance) {
+    private int getPortion(double distance, int additionalFareDistance, int additionalStandardDistance) {
         int portion = (int) (distance - additionalFareDistance) / additionalStandardDistance;
         if (portion == 0) {
-            portion = 1;
+            portion = DEFAULT_PORTION;
         }
         return portion;
     }
 
-    private static int calculateFare(int portion) {
+    private int calculateFare(int portion) {
         return portion * MINIMUM_ADDITIONAL_FARE;
     }
 }
