+++ src/main/java/wooteco/subway/domain/fare/strategy/FareStrategy.java
@@ -1,0 +1,6 @@
+package wooteco.subway.domain.fare.strategy;
+
+public interface FareStrategy {
+
+    int calculate(double distance);
+}
