+++ src/main/java/wooteco/subway/service/PathService.java
@@ -5,10 +5,11 @@ import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
+import wooteco.subway.domain.fare.FareCalculator;
+import wooteco.subway.domain.fare.strategy.DefaultFareStrategy;
 import wooteco.subway.domain.path.Path;
 import wooteco.subway.domain.path.PathFinder;
 import wooteco.subway.dto.PathResponse;
-import wooteco.subway.util.FareCalculator;
 
 @Service
 public class PathService {
@@ -25,8 +26,9 @@ public class PathService {
         PathFinder pathFinder = new PathFinder();
         Path path = pathFinder.getShortestPath(getStationById(sourceId), getStationById(targetId),
                 new Sections(sectionDao.findAll()));
+        FareCalculator fareCalculator = new FareCalculator(new DefaultFareStrategy());
 
-        return PathResponse.from(path, FareCalculator.calculate(path.getDistance()));
+        return PathResponse.from(path, fareCalculator.calculate(path.getDistance()));
     }
 
     private Station getStationById(Long id) {
